+++ includes/footer.php
@@ -1,6 +1,6 @@
 <div class="container">
   <footer class="footer">
-    <p>Copyright &copy; 2015-2016 Cyotek Ltd. All Rights Reserved.</p>
+    <p>Copyright &copy; 2015-2017 Cyotek Ltd. All Rights Reserved.</p>
     <p><a href="mailto:support@cyotek.com?subject=<?php echo str_replace('+', '%20', urlencode(SITE_TITLE . ' Feedback (' . $_SERVER["REQUEST_URI"] . ')')) ?>">Comments or suggestions?</a></p>
     <p class="version">Website version <?php echo WEBSITE_VERSION ?></p>
   </footer>