+++ javascript/1.index.php
@@ -1,0 +1,13 @@
+<?php require_once __DIR__ . '/prepend.php'; ?>
+
+<?php 
+
+  define('PAGE_TITLE', 'Category 1 Index');
+
+  renderJavaScriptTemplateStart(); 
+ 
+?>
+
+  <p>This page serves no purpose other than being the destination from a JavaScript function.</p>
+
+<?php renderJavaScriptTemplateEnd(); ?>